replace gp with sklearn_gp (#14)
